requirements.txt
anyio
mcp
pydantic
pytest
numpy
faiss-cpu
